Thêm lớp character để hiển thị điểm khi giết quái (hiện tại chỉ có số). Điều chỉnh các lớp khác để thêm thuộc tính character. Thêm biến score lưu điểm cho mario. Cho phép reload scene hiện tại (ấn R).

+++ 05-SceneManager/ShootingPlant.h
@@ -23,7 +23,7 @@ private:
     float detectRange;
 
     float startY, topY;
-    DWORD stateTimer;
+    ULONGLONG stateTimer;
     SPLANT_STATE state;
     bool hasShot;
 public:
